# Switched to rent an array in Giraffe handler