Looking at allocations seems like an improvement.
Probably the time is not comparable if my machine was loaded differently during runs.

New
| Method                         | Mean      | Error     | StdDev    | Median    | Min       | Max       | Op/s     | Gen0    | Gen1   | Allocated |
|------------------------------- |----------:|----------:|----------:|----------:|----------:|----------:|---------:|--------:|-------:|----------:|
| BenchmarkSimpleQueryUnparsed   |  93.97 us |  0.984 us |  0.921 us |  93.69 us |  92.67 us |  95.34 us | 10,641.6 |  2.3193 |      - |     19 KB |
| BenchmarkSimpleQueryParsed     | 114.89 us |  5.955 us | 17.466 us | 119.22 us |  85.92 us | 150.19 us |  8,703.9 |  1.0986 |      - |   9.56 KB |
| BenchmarkSimpleQueryPlanned    |  43.56 us |  2.285 us |  6.665 us |  42.20 us |  33.68 us |  60.05 us | 22,957.3 |  0.9155 |      - |   7.58 KB |
| BenchmarkFlatQueryUnparsed     | 296.69 us | 10.119 us | 29.034 us | 285.64 us | 261.22 us | 375.81 us |  3,370.5 |  7.3242 |      - |  63.06 KB |
| BenchmarkFlatQueryParsed       | 246.00 us |  4.903 us | 13.255 us | 241.75 us | 230.02 us | 287.13 us |  4,065.1 |  5.3711 |      - |  47.15 KB |
| BenchmarkFlatQueryPlanned      | 180.46 us |  3.587 us |  7.164 us | 179.69 us | 170.01 us | 198.97 us |  5,541.4 |  5.1270 | 0.2441 |  43.77 KB |
| BenchmarkNestedQueryUnparsed   | 438.59 us |  4.732 us |  4.195 us | 438.19 us | 432.38 us | 446.10 us |  2,280.0 | 20.5078 | 0.4883 | 169.34 KB |
| BenchmarkNestedQueryParsed     | 393.96 us |  7.863 us | 13.563 us | 387.65 us | 377.57 us | 425.99 us |  2,538.3 | 14.6484 | 0.4883 | 120.07 KB |
| BenchmarkNestedQueryPlanned    | 344.46 us |  6.842 us | 17.292 us | 340.46 us | 315.08 us | 394.41 us |  2,903.1 | 13.1836 |      - | 109.41 KB |
| BenchmarkFilteredQueryUnparsed | 198.84 us |  3.941 us |  7.206 us | 196.93 us | 188.97 us | 218.31 us |  5,029.1 |  6.8359 |      - |  56.48 KB |
| BenchmarkFilteredQueryParsed   | 158.51 us |  3.165 us |  3.767 us | 157.44 us | 152.62 us | 166.04 us |  6,308.8 |  3.4180 |      - |  29.16 KB |
| BenchmarkFilteredQueryPlanned  | 117.62 us |  4.301 us | 12.271 us | 114.44 us |  98.69 us | 153.09 us |  8,502.1 |  2.8076 |      - |  23.67 KB |

Old
| Method                         | Mean      | Error    | StdDev   | Min       | Max       | Op/s     | Gen0    | Gen1   | Allocated |
|------------------------------- |----------:|---------:|---------:|----------:|----------:|---------:|--------:|-------:|----------:|
| BenchmarkSimpleQueryUnparsed   |  83.25 us | 1.649 us | 1.543 us |  81.07 us |  87.04 us | 12,011.6 |  2.3193 |      - |     19 KB |
| BenchmarkSimpleQueryParsed     |  68.27 us | 1.306 us | 1.282 us |  66.59 us |  70.87 us | 14,647.6 |  1.0986 |      - |   9.56 KB |
| BenchmarkSimpleQueryPlanned    |  29.32 us | 0.535 us | 0.418 us |  28.60 us |  30.07 us | 34,111.6 |  0.9155 |      - |   7.58 KB |
| BenchmarkFlatQueryUnparsed     | 233.03 us | 2.654 us | 2.482 us | 228.53 us | 235.83 us |  4,291.2 |  7.3242 |      - |  63.07 KB |
| BenchmarkFlatQueryParsed       | 210.31 us | 2.704 us | 2.529 us | 206.20 us | 214.19 us |  4,755.0 |  5.6152 | 0.2441 |  47.11 KB |
| BenchmarkFlatQueryPlanned      | 162.98 us | 1.722 us | 1.610 us | 161.00 us | 165.93 us |  6,135.8 |  5.1270 | 0.2441 |  43.77 KB |
| BenchmarkNestedQueryUnparsed   | 398.08 us | 2.335 us | 1.950 us | 394.31 us | 401.22 us |  2,512.1 | 20.5078 | 0.4883 | 169.34 KB |
| BenchmarkNestedQueryParsed     | 350.29 us | 3.153 us | 2.633 us | 347.09 us | 356.82 us |  2,854.8 | 14.6484 | 0.4883 | 120.07 KB |
| BenchmarkNestedQueryPlanned    | 299.97 us | 2.615 us | 2.184 us | 295.74 us | 303.87 us |  3,333.7 | 13.1836 |      - | 109.41 KB |
| BenchmarkFilteredQueryUnparsed | 173.42 us | 2.713 us | 2.538 us | 169.62 us | 177.50 us |  5,766.2 |  6.8359 |      - |  56.48 KB |
| BenchmarkFilteredQueryParsed   | 140.57 us | 2.723 us | 2.913 us | 135.77 us | 146.48 us |  7,114.0 |  3.4180 |      - |  29.16 KB |
| BenchmarkFilteredQueryPlanned  |  92.81 us | 1.503 us | 1.332 us |  90.71 us |  95.11 us | 10,774.7 |  2.8076 |      - |  23.67 KB |